requirements.txt
streamlit
langchain
python-pptx
Pillow
pytesseract
qdrant-client
fastembed
